+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.9.0-SNAPSHOT] - 2020-05-09
+### Added
+- A deleted flag to price entity
+- Deleting a price entry (by setting its deleted flag to true)
+- Listing all deleted prices
+
+### Changed
+- List all prices: all -> not deleted prices
+- Restricted updating a price to only non-deleted prices
+
 ## [1.8.0-SNAPSHOT] - 2020-05-02
 ### Added
 - Updating a price entry